add routing for the API
