+++ .env
@@ -10,3 +10,7 @@ COMPUTE_TYPE="int8_float16"
 NEMO_DOMAIN_TYPE="telephonic"  # Can be general, meeting or telephonic based on domain type of the audio file
 NEMO_STORAGE_PATH="nemo_storage"
 NEMO_OUTPUT_PATH="nemo_outputs"
+AUDIO_FILE_ENDPOINT=True
+AUDIO_URL_ENDPOINT=True
+YOUTUBE_ENDPOINT=True
+LIVE_ENDPOINT=False
