+++ wordcab_transcribe/main.py
@@ -15,28 +15,15 @@
 """Main API module of the Wordcab Transcribe."""
 
 import asyncio
-from typing import Optional
 
-import aiofiles
-import shortuuid
-from fastapi import BackgroundTasks, FastAPI, File, UploadFile
-from fastapi import status as http_status
+from fastapi import FastAPI
 from fastapi.responses import HTMLResponse
 from loguru import logger
 
 from wordcab_transcribe.config import settings
-from wordcab_transcribe.models import ASRResponse, DataRequest
-from wordcab_transcribe.service import ASRService
-from wordcab_transcribe.utils import (
-    convert_file_to_wav,
-    convert_timestamp,
-    delete_file,
-    download_audio_file,
-    download_file_from_youtube,
-    format_punct,
-    is_empty_string,
-    retrieve_user_platform,
-)
+from wordcab_transcribe.dependencies import asr
+from wordcab_transcribe.router.v1.endpoints import api_router
+from wordcab_transcribe.utils import retrieve_user_platform
 
 
 app = FastAPI(
@@ -46,7 +33,7 @@ app = FastAPI(
     debug=settings.debug,
 )
 
-asr = ASRService()
+app.include_router(api_router, prefix=settings.api_prefix)
 
 
 @app.on_event("startup")
@@ -92,137 +79,3 @@ async def health_check():
     </html>
     """
     return HTMLResponse(content=content, media_type="text/html")
-
-
-@app.post(
-    f"{settings.api_prefix}/audio",
-    tags=["inference"],
-    response_model=ASRResponse,
-    status_code=http_status.HTTP_200_OK,
-)
-async def inference_with_audio(
-    background_tasks: BackgroundTasks,
-    file: UploadFile = File(...),  # noqa: B008
-    data: Optional[DataRequest] = None,
-) -> ASRResponse:
-    """Inference endpoint with audio file."""
-    extension = file.filename.split(".")[-1]
-    filename = f"audio_{shortuuid.ShortUUID().random(length=32)}.{extension}"
-
-    async with aiofiles.open(filename, "wb") as f:
-        audio_bytes = await file.read()
-        await f.write(audio_bytes)
-
-    if extension != "wav":
-        filepath = await convert_file_to_wav(filename)
-        background_tasks.add_task(delete_file, filepath=filename)
-    else:
-        filepath = filename
-
-    if data is None:
-        data = DataRequest()
-    else:
-        data = DataRequest(**data.dict())
-
-    raw_utterances = await asr.process_input(filepath, data.source_lang)
-
-    timestamps_format = data.timestamps
-    utterances = [
-        {
-            "text": format_punct(utterance["text"]),
-            "start": convert_timestamp(utterance["start"], timestamps_format),
-            "end": convert_timestamp(utterance["end"], timestamps_format),
-            "speaker": int(utterance["speaker"]),
-        }
-        for utterance in raw_utterances
-        if not is_empty_string(utterance["text"])
-    ]
-
-    background_tasks.add_task(delete_file, filepath=filepath)
-
-    return ASRResponse(utterances=utterances)
-
-
-@app.post(
-    f"{settings.api_prefix}/youtube",
-    tags=["inference"],
-    response_model=ASRResponse,
-    status_code=http_status.HTTP_200_OK,
-)
-async def inference_with_youtube(
-    background_tasks: BackgroundTasks,
-    url: str,
-    data: Optional[DataRequest] = None,
-) -> ASRResponse:
-    """Inference endpoint with YouTube url."""
-    filename = f"yt_{shortuuid.ShortUUID().random(length=32)}"
-    filepath = await download_file_from_youtube(url, filename)
-
-    if data is None:
-        data = DataRequest()
-    else:
-        data = DataRequest(**data.dict())
-
-    raw_utterances = await asr.process_input(filepath, data.source_lang)
-
-    timestamps_format = data.timestamps
-    utterances = [
-        {
-            "text": format_punct(utterance["text"]),
-            "start": convert_timestamp(utterance["start"], timestamps_format),
-            "end": convert_timestamp(utterance["end"], timestamps_format),
-            "speaker": int(utterance["speaker"]),
-        }
-        for utterance in raw_utterances
-        if not is_empty_string(utterance["text"])
-    ]
-
-    background_tasks.add_task(delete_file, filepath=filepath)
-
-    return ASRResponse(utterances=utterances)
-
-
-@app.post(
-    f"{settings.api_prefix}/audio-url",
-    tags=["inference"],
-    response_model=ASRResponse,
-    status_code=http_status.HTTP_200_OK,
-)
-async def inference_with_audio_url(
-    background_tasks: BackgroundTasks,
-    url: str,
-    data: Optional[DataRequest] = None,
-) -> ASRResponse:
-    """Inference endpoint with audio url."""
-    filename = f"audio_url_{shortuuid.ShortUUID().random(length=32)}"
-    filepath = await download_audio_file(url, filename)
-    extension = filepath.split(".")[-1]
-
-    if extension != "wav":
-        filepath = await convert_file_to_wav(filepath)
-        background_tasks.add_task(delete_file, filepath=f"{filename}.{extension}")
-    else:
-        filepath = filename
-
-    if data is None:
-        data = DataRequest()
-    else:
-        data = DataRequest(**data.dict())
-
-    raw_utterances = await asr.process_input(filepath, data.source_lang)
-
-    timestamps_format = data.timestamps
-    utterances = [
-        {
-            "text": format_punct(utterance["text"]),
-            "start": convert_timestamp(utterance["start"], timestamps_format),
-            "end": convert_timestamp(utterance["end"], timestamps_format),
-            "speaker": int(utterance["speaker"]),
-        }
-        for utterance in raw_utterances
-        if not is_empty_string(utterance["text"])
-    ]
-
-    background_tasks.add_task(delete_file, filepath=filepath)
-
-    return ASRResponse(utterances=utterances)
