+++ wordcab_transcribe/router/v1/live_endpoints.py
@@ -1,0 +1,25 @@
+# Copyright 2023 The Wordcab Team. All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""Live endpoints for the Wordcab Transcribe API."""
+
+from fastapi import APIRouter
+
+
+router = APIRouter()
+
+
+@router.post("/live")
+async def live() -> None:
+    """Live endpoint for the API."""
+    pass
